test(scan): add scalar, array micro perf test for scan operator
- remove perf test for scan under current thread scheduler
